#5893: Update test files
