+++ tests/ttnn/unit_tests/operations/eltwise/test_eltwise_rsqrt.py
@@ -9,11 +9,11 @@ import torch
 import ttnn
 import traceback
 
-from tests.ttnn.utils_for_testing import assert_with_pcc
 from tests.ttnn.python_api_testing.sweep_tests import ttnn_ops
+from tests.tt_eager.python_api_testing.sweep_tests.comparison_funcs import comp_pcc
 
 
-def run_eltwise_sin_tests(
+def run_eltwise_rsqrt_tests(
     input_shape,
     dtype,
     dlayout,
@@ -23,16 +23,18 @@ def run_eltwise_sin_tests(
     device,
 ):
     torch.manual_seed(data_seed)
-    x = torch.Tensor(size=input_shape[0]).uniform_(-100, 100).to(torch.bfloat16)
+    x = torch.Tensor(size=input_shape[0]).uniform_(1, 100).to(torch.bfloat16)  # Range will be updated after #14077.
 
     try:
         # get ref result
-        ref_value = torch.sin(x)
+        ref_value = torch.rsqrt(x)
+        logger.info(f"PyTorch: {ref_value[0:10, 0:10]}")
 
         x = ttnn_ops.setup_ttnn_tensor(x, device, dlayout[0], in_mem_config[0], dtype[0])
 
-        tt_result = ttnn.sin(x)
+        tt_result = ttnn.rsqrt(x)
         tt_result = ttnn_ops.ttnn_tensor_to_torch(tt_result, output_mem_config)
+        logger.info(f"TensTorrent: {tt_result[0:10, 0:10]}")
 
     except Exception as e:
         logger.warning(f"Test execution crashed: {e}")
@@ -41,25 +43,31 @@ def run_eltwise_sin_tests(
 
     assert len(tt_result.shape) == len(ref_value.shape)
     assert tt_result.shape == ref_value.shape
-    assert_with_pcc(ref_value, tt_result, 0.99)
+
+    # compare tt and golden outputs
+    success, pcc_value = comp_pcc(ref_value, tt_result)
+    logger.debug(pcc_value)
+    logger.debug(success)
+
+    assert success
 
 
 test_sweep_args = [
     (
-        [(3, 2, 192, 32)],
-        [ttnn.bfloat16],
+        [(224, 128)],
+        [ttnn.bfloat8_b],
         [ttnn.TILE_LAYOUT],
         [ttnn.DRAM_MEMORY_CONFIG],
         ttnn.DRAM_MEMORY_CONFIG,
-        11079580,
+        1358430,
     ),
     (
-        [(12, 224, 224)],
+        [(6, 160, 64)],
         [ttnn.bfloat8_b],
         [ttnn.TILE_LAYOUT],
         [ttnn.DRAM_MEMORY_CONFIG],
         ttnn.L1_MEMORY_CONFIG,
-        6411147,
+        17869870,
     ),
 ]
 
@@ -68,5 +76,5 @@ test_sweep_args = [
     "input_shape, dtype, dlayout, in_mem_config, out_mem_config, data_seed",
     (test_sweep_args),
 )
-def test_eltwise_sin(input_shape, dtype, dlayout, in_mem_config, out_mem_config, data_seed, device):
-    run_eltwise_sin_tests(input_shape, dtype, dlayout, in_mem_config, out_mem_config, data_seed, device)
+def test_eltwise_rsqrt(input_shape, dtype, dlayout, in_mem_config, out_mem_config, data_seed, device):
+    run_eltwise_rsqrt_tests(input_shape, dtype, dlayout, in_mem_config, out_mem_config, data_seed, device)
